Extracted code from UnpublishAllArticlesWorker to the service
